Split ShareContainer into 2 separate components, 1 component to email takes and another to display a dropdown menu when clicked.
