+++ client/src/main/scripts/components/ReactionContainer.tsx
@@ -5,7 +5,8 @@ import { TakeReactionJson } from "../java2ts/TakeReactionJson";
 import { postRequest } from "../utils/databaseAPI";
 import { alertErr } from "../utils/functions";
 import { TakeDocument } from "./BlockEditor";
-import ShareContainer from "./ShareContainer";
+import DropDown from "./DropDown";
+import EmailTake from "./EmailTake";
 
 interface ReactionContainerProps {
   takeId: number;
@@ -235,12 +236,14 @@ export class Reaction extends React.Component<ReactionProps, ReactionState> {
               <i className="fa fa-star-o" aria-hidden="true" />
             )}
           </button>
-          <ShareContainer
-            takeDocument={(Object as any).assign(
-              {},
-              props.containerProps.takeDocument
-            )}
-          />
+          <DropDown text="Email" position="TL">
+            <EmailTake
+              takeDocument={(Object as any).assign(
+                {},
+                props.containerProps.takeDocument
+              )}
+            />
+          </DropDown>
           <div className="reaction__action-container">
             <button
               className="reaction__action reaction__action--report"
